Split out query parsing into its own class, and start adding unit tests for queries
Move parsing from a Query.from_args class method to its own QueryParser class.
Introduce an in-memory DictSource, and use it to build unit tests for queries.
Start with an aggregated query to exercise the sqlite parts of the system.
Allow for data sources to be be passed directly to the parser, as well as just strings,
to make it easier to send pre-canned data for self-testing purposes.